Refactored muEd into evaluate and chat
